+++ flask_unchained/unchained.py
@@ -14,6 +14,90 @@ from .exceptions import ServiceUsageError
 from .utils import AttrDict
 
 
+class _DeferredBundleFunctions:
+    def __init__(self):
+        self._deferred_functions = []
+
+    def _defer(self, fn):
+        self._deferred_functions.append(fn)
+
+    def before_request(self, fn):
+        """
+        Like :meth:`~flask.Blueprint.before_request` but for a bundle. This function
+        is only executed before each request that is handled by a view function
+        of that bundle.
+        """
+        self._defer(lambda bp: bp.before_request(fn))
+
+    def after_request(self, fn):
+        """
+        Like :meth:`~flask.Blueprint.after_request` but for a bundle. This function
+        is only executed after each request that is handled by a function of
+        that bundle.
+        """
+        self._defer(lambda bp: bp.after_request(fn))
+
+    def teardown_request(self, fn):
+        """
+        Like :meth:`~flask.Blueprint.teardown_request` but for a bundle. This
+        function is only executed when tearing down requests handled by a
+        function of that bundle.  Teardown request functions are executed
+        when the request context is popped, even when no actual request was
+        performed.
+        """
+        self._defer(lambda bp: bp.teardown_request(fn))
+
+    def context_processor(self, fn):
+        """
+        Like :meth:`~flask.Blueprint.context_processor` but for a bundle. This
+        function is only executed for requests handled by a bundle.
+        """
+        self._defer(lambda bp: bp.context_processor(fn))
+        return fn
+
+    def url_defaults(self, fn):
+        """
+        Callback function for URL defaults for this bundle. It's called
+        with the endpoint and values and should update the values passed
+        in place.
+        """
+        self._defer(lambda bp: bp.url_defaults(fn))
+        return fn
+
+    def url_value_preprocessor(self, fn):
+        """
+        Registers a function as URL value preprocessor for this
+        bundle. It's called before the view functions are called and
+        can modify the url values provided.
+        """
+        self._defer(lambda bp: bp.url_value_preprocessor(fn))
+        return fn
+
+    def errorhandler(self, code_or_exception):
+        """
+        Registers an error handler that becomes active for this bundle
+        only.  Please be aware that routing does not happen local to a
+        bundle so an error handler for 404 usually is not handled by
+        a bundle unless it is caused inside a view function.  Another
+        special case is the 500 internal server error which is always looked
+        up from the application.
+
+        Otherwise works as the :meth:`~flask.Blueprint.errorhandler` decorator.
+        """
+        def decorator(fn):
+            self._defer(lambda bp: bp.register_error_handler(code_or_exception, fn))
+            return fn
+        return decorator
+
+
+class _DeferredBundleFunctionsStore:
+    def __init__(self):
+        self._bundles = {}
+
+    def __getitem__(self, bundle_name):
+        if bundle_name not in self._bundles:
+            self._bundles[bundle_name] = _DeferredBundleFunctions()
+        return self._bundles[bundle_name]
 
 
 class Unchained:
@@ -27,6 +111,8 @@ class Unchained:
 
     def __init__(self, env: Optional[Union[DEV, PROD, STAGING, TEST]] = None):
         self.bundles = AttrDict()
+        self._bundles = _DeferredBundleFunctionsStore()
+        self.app_bundle = None
         self.babel_bundle = None
         self.env = env
         self.extensions = AttrDict()
@@ -46,14 +132,18 @@ class Unchained:
                  bundles: Optional[List] = None,
                  _config_overrides: Optional[Dict[str, Any]] = None,
                  ) -> None:
-        self.bundles = (AttrDict({b.name: b for b in bundles})
-                        if bundles else AttrDict())
-        self.env = env or self.env
-        app.env = self.env
-        app.extensions['unchained'] = self
-        app.unchained = self
-        app.shell_context_processor(lambda: {
-            bundle.__class__.__name__: bundle for bundle in bundles})
+        bundles = bundles or []
+        for b in bundles:
+            b._deferred_functions = self._bundles[b.name]._deferred_functions
+        self.bundles = AttrDict({b.name: b for b in bundles})
+        app.shell_context_processor(lambda: {b.__class__.__name__: b for b in bundles})
+
+        try:
+            # must import AppBundle here to prevent circular dependency
+            from .bundle import AppBundle
+            self.app_bundle = [b for b in bundles if isinstance(b, AppBundle)][0]
+        except IndexError:
+            self.app_bundle = None
 
         try:
             # must import BabelBundle here to prevent circular dependency
@@ -61,6 +151,12 @@ class Unchained:
             self.babel_bundle = [b for b in bundles if isinstance(b, BabelBundle)][0]
         except IndexError:
             self.babel_bundle = None
+
+        self.env = env or self.env
+        app.env = self.env
+
+        app.extensions['unchained'] = self
+        app.unchained = self
 
         for deferred in self._deferred_functions:
             deferred(app)
@@ -545,6 +641,8 @@ class Unchained:
     def __getattr__(self, name: str):
         if name in self.bundles:
             return self.bundles[name]
+        elif not self._initialized:
+            return self._bundles[name]
         raise AttributeError(name)